Improve install script
Allows installation globally and in an existing virtual environment,
though the latter requires confirmation from the user.
Downloading the depccg pre-trained parser is now an interactive prompt.

Furthermore, restricting the version of `chainer`, which is required for
`pip` to behave, also fixes the `typing-extensions` issue, so that hack
has been removed.