Rewrite CCG rule tests
Since all the rule tests share the same structure, this formalises that
by defining a base class `CCGRuleTester`, which groups all the test
functions in a class. Each rule subclass thus defines the test tree and
the expected diagrams to match. This reduces the size of the tests
enough to allow combining them into a single file.
Furthermore, the `CCGTree` parameter `ccg_rule` is renamed to `rule`.