Make depccg an optional dependency and update the installer.
This allows Lambeq to be installed on Windows, and makes it directly
pip-installable, which is a better user experience.
The installer has been made more interactive and can be used to install
one of multiple configurations of dependencies; it also detects whether
the user is already in a virtual environment so can be directly
installed there.
The README has also been updated to reflect these changes.