tests: improve slow-running tests by mocking and other tricks.
One way is to change the root categories in Bobcat directly rather than
creating a new object. To do this, the function
`ChartParser.set_root_cats` was added. This functionality could be
exposed in `BobcatParser` in a nice way to allow users to change root
categories on the fly.

Also, this adds a line to the workflow to print out the 50
slowest-running tests.