workflows: refactor depccg-related tests
A depccg dependency, allennlp, requires `torch<1.12`, which conflicts
with lambeq. To fix this, we now re-install lambeq after installing
depccg in the tests. As well as this, we also move the depccg-related
tests to the end, so they are all done in one block.
Also, we re-enable depccg tests for 3.10, since it appears to work now.