Add option to maintain or collapse noun phrases as desired (#590)
This adds the parameter `collapse_noun_phrases` to sentence-to-diagram
methods of `CCGParser` and `TreeReader` which controls whether NP types
are changed into N types. This is set to `True` by default for backwards
compatibility.

This is only done at the tree-to-diagram step, so this has the breaking
change of CCG trees containing NP types by default. In order to change
them to N types, call `.collapse_noun_phrases()` on each tree.

Co-authored-by: Dimitri Kartsaklis <dimkart@gmail.com>